+++ hostlib/hostLib/inc/PlugAndTrust_Pkg_Ver.h
@@ -11,18 +11,18 @@
 
 /* clang-format off */
 #define PLUGANDTRUST_PROD_NAME          "PlugAndTrust"
-#define PLUGANDTRUST_VER_STRING_NUM     "v04.07.00_20250210"
-#define PLUGANDTRUST_PROD_NAME_VER_FULL "PlugAndTrust_v04.07.00_20250210"
+#define PLUGANDTRUST_VER_STRING_NUM     "v04.07.01_20250519"
+#define PLUGANDTRUST_PROD_NAME_VER_FULL "PlugAndTrust_v04.07.01_20250519"
 #define PLUGANDTRUST_VER_MAJOR          (4u)
 #define PLUGANDTRUST_VER_MINOR          (7u)
-#define PLUGANDTRUST_VER_DEV            (0u)
+#define PLUGANDTRUST_VER_DEV            (1u)
 
 /* v04.07 = 40007u */
 #define PLUGANDTRUST_VER_MAJOR_MINOR ( 0 \
     | (PLUGANDTRUST_VER_MAJOR * 10000u)    \
     | (PLUGANDTRUST_VER_MINOR))
 
-/* v04.07.00 = 400070000ULL */
+/* v04.07.01 = 400070001ULL */
 #define PLUGANDTRUST_VER_MAJOR_MINOR_DEV ( 0 \
     | (PLUGANDTRUST_VER_MAJOR * 10000*10000u)    \
     | (PLUGANDTRUST_VER_MINOR * 10000u)    \
@@ -50,7 +50,7 @@
  *
  * v_minor  = "07"
  *
- * v_dev    = "00"
+ * v_dev    = "01"
  *
  * # Develop Branch
  * v_meta   = ""