+++ sss/inc/fsl_sss_api_ver.h
@@ -10,18 +10,18 @@
 
 /* clang-format off */
 #define SSS_APIS_PROD_NAME          "SSS_APIs"
-#define SSS_APIS_VER_STRING_NUM     "v04.07.00_20250210"
-#define SSS_APIS_PROD_NAME_VER_FULL "SSS_APIs_v04.07.00_20250210"
+#define SSS_APIS_VER_STRING_NUM     "v04.07.01_20250519"
+#define SSS_APIS_PROD_NAME_VER_FULL "SSS_APIs_v04.07.01_20250519"
 #define SSS_APIS_VER_MAJOR          (4u)
 #define SSS_APIS_VER_MINOR          (7u)
-#define SSS_APIS_VER_DEV            (0u)
+#define SSS_APIS_VER_DEV            (1u)
 
 /* v04.07 = 40007u */
 #define SSS_APIS_VER_MAJOR_MINOR ( 0 \
     | (SSS_APIS_VER_MAJOR * 10000u)    \
     | (SSS_APIS_VER_MINOR))
 
-/* v04.07.00 = 400070000ULL */
+/* v04.07.01 = 400070001ULL */
 #define SSS_APIS_VER_MAJOR_MINOR_DEV ( 0 \
     | (SSS_APIS_VER_MAJOR * 10000*10000u)    \
     | (SSS_APIS_VER_MINOR * 10000u)    \
@@ -48,7 +48,7 @@
  *
  * v_minor  = "07"
  *
- * v_dev    = "00"
+ * v_dev    = "01"
  *
  * v_meta   = ""
  *